+++ resources/pod-emacs-ob-babashka.el
@@ -1,0 +1,6 @@
+;; pod-emacs-ob-babashka.el --- support for running bb in org-babel -*- lexical-binding:t; -*-
+
+(pod-emacs-register
+ "pod.kpassapk.emacs.ob-babashka" '())
+
+(provide 'pod-emacs-ob-babashka)